fix: Model 내 프로퍼티의 잘못된 타입 수정
* WeatherInformation, FiveDayData 타입 내의 프로퍼티 수정

+++ WeatherForecast/WeatherForecast/Model/WeatherInformation.swift
@@ -63,8 +63,8 @@ struct Sys: Codable {
 }
 
 struct Rain: Codable {
-    let oneHour: Int?
-    let threeHours: Int?
+    let oneHour: Double?
+    let threeHours: Double?
     
     enum CodingKeys: String, CodingKey {
         case oneHour = "1h"
@@ -73,8 +73,8 @@ struct Rain: Codable {
 }
 
 struct Snow: Codable {
-    let oneHour: Int?
-    let threeHours: Int?
+    let oneHour: Double?
+    let threeHours: Double?
     
     enum CodingKeys: String, CodingKey {
         case oneHour = "1h"
